chore: Rename utils.py as interop.py
